v4.2.2
.- Usuario perfil Jefe Tecnico Nacional sólo debe ver las solicitudes de vacaciones de empleados con cargo ID=22 (Director)

.- Se modifica la función 'setsaldodiasvacacionesasignadas':
Arrojaba un error de parseo y posterior redondeo de decimales.

.- Vista de 'Admin Vacaciones': Ahora se muestran los decimales en las columnas: "Dias acum. vac asignadas" y "Saldo vac asignadas".

+++ GestionAsistenciaFemase/src/java/cl/femase/gestionweb/dao/DetalleAusenciaDAO.java
@@ -727,8 +727,8 @@ public class DetalleAusenciaDAO extends BaseDAO{
                 data.setEsZonaExtrema(rs.getString("es_zona_extrema"));
                 
                 //nuevas columnas
-                data.setSaldoDiasVacacionesAsignadas(rs.getInt("saldo_dias_vacaciones_asignadas"));
-                data.setDiasAcumuladosVacacionesAsignadas(rs.getInt("dias_acumulados_vacaciones_asignadas"));        
+                data.setSaldoDiasVacacionesAsignadas(rs.getDouble("saldo_dias_vacaciones_asignadas"));
+                data.setDiasAcumuladosVacacionesAsignadas(rs.getDouble("dias_acumulados_vacaciones_asignadas"));        
     
                 lista.add(data);
             }
@@ -1416,8 +1416,9 @@ public class DetalleAusenciaDAO extends BaseDAO{
                 data.setIdAusencia(rs.getInt("ausencia_id"));
                 data.setEmpresaId(rs.getString("empresa_id"));
                 data.setDiasEfectivosVacaciones(rs.getInt("dias_efectivos_vacaciones"));
-                data.setSaldoDiasVacacionesAsignadas(rs.getInt("saldo_dias_vacaciones_asignadas"));
-                data.setDiasAcumuladosVacacionesAsignadas(rs.getInt("dias_acumulados_vacaciones_asignadas"));
+                //nuevas by Ramon
+                data.setSaldoDiasVacacionesAsignadas(rs.getDouble("saldo_dias_vacaciones_asignadas"));
+                data.setDiasAcumuladosVacacionesAsignadas(rs.getDouble("dias_acumulados_vacaciones_asignadas"));
             }
 
             ps.close();
